+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Video: Makes `aspectRatio` a required prop for `Video` (#201)
 * Video: Pass events through to callback functions (#203)
 * Touchable: Add event targets to Flow typing for callbacks (#204)
+* Video: Add new `onEnded` prop for media end event (#207)
 
 ### Patch
 * Internal: Add code coverage to PRs (#185)